making CommandPalette public again, creating trivial wrapper for plugin
