+++ src/commandpalette/palette.css
